rename img_spec to img_gen_spec
