+++ pipelex/libraries/pipelines/builder/pipe/pipe_img_gen_spec.py
