+++ tests/unit/pipelex/libraries/pipelines/builder/pipe/pipe_operator/pipe_img_gen/test_data.py
@@ -1,7 +1,7 @@
 from typing import ClassVar
 
 from pipelex.core.pipes.input_requirement_blueprint import InputRequirementBlueprint
-from pipelex.libraries.pipelines.builder.pipe.pipe_img_spec import ImgGenSkill, PipeImgGenSpec
+from pipelex.libraries.pipelines.builder.pipe.pipe_img_gen_spec import ImgGenSkill, PipeImgGenSpec
 from pipelex.pipe_operators.img_gen.pipe_img_gen_blueprint import PipeImgGenBlueprint
