+++ tests/unit/pipelex/libraries/pipelines/builder/pipe/pipe_operator/pipe_img_gen/test_pipe_img_gen.py
@@ -1,6 +1,6 @@
 import pytest
 
-from pipelex.libraries.pipelines.builder.pipe.pipe_img_spec import PipeImgGenSpec
+from pipelex.libraries.pipelines.builder.pipe.pipe_img_gen_spec import PipeImgGenSpec
 from pipelex.pipe_operators.img_gen.pipe_img_gen_blueprint import PipeImgGenBlueprint
 from tests.unit.pipelex.libraries.pipelines.builder.pipe.pipe_operator.pipe_img_gen.test_data import PipeImgGenTestCases
